refactor: remove unused sessionmaker and database connection management
