+++ app/session.py
@@ -1,0 +1,18 @@
+from sqlalchemy.orm import sessionmaker
+
+from app.database import engine
+
+
+SessionLocal = sessionmaker(
+    autocommit=False,
+    autoflush=False,
+    bind=engine
+)
+
+def get_db():
+    db = SessionLocal()
+
+    try:
+        yield db
+    finally:
+        db.close()
